Change function to make it generic
